+++ spec/editor-spec.js
@@ -4,9 +4,9 @@
 
 /* eslint-disable prefer-const */
 
-import { Range } from 'atom'
+import { Range, Point } from 'atom'
 import Editor from '../lib/editor'
-import { getMessage } from './helpers'
+import { it, wait, getMessage, generateEvent } from './helpers'
 
 describe('Editor', function() {
   let editor
@@ -47,6 +47,34 @@ describe('Editor', function() {
       editor.apply([], [message])
       expect(Range.fromObject(message.range)).toEqual({ start: { row: 2, column: 0 }, end: { row: 2, column: 6 } })
       expect(textEditor.getMarkerCount()).toBe(0)
+    })
+  })
+  describe('Response to config', function() {
+    it('responds to `tooltipFollows` config', async function() {
+      const position = [2, 1]
+      const editorElement = atom.views.getView(textEditor)
+      editor.apply([
+        getMessage('Error', __filename, Range.fromObject([position, [Infinity, Infinity]]))
+      ], [])
+
+      atom.config.set('linter-ui-default.tooltipFollows', 'Keyboard')
+      expect(editor.bubble).toBe(null)
+      textEditor.setCursorBufferPosition(position)
+      await wait(60)
+      expect(editor.bubble).not.toBe(null)
+      expect(typeof editor.bubble.destroy).toBe('function')
+
+      atom.config.set('linter-ui-default.tooltipFollows', 'Mouse')
+      expect(editor.bubble).toBe(null)
+      const event = generateEvent(editorElement, 'mousemove')
+      Object.defineProperty(event, 'target', { value: editorElement })
+      editorElement.dispatchEvent(event)
+      spyOn(textEditor, 'bufferPositionForScreenPosition').andCallFake(function() {
+        return Point.fromObject(position)
+      })
+      await wait(200)
+      expect(editor.bubble).not.toBe(null)
+      expect(typeof editor.bubble.destroy).toBe('function')
     })
   })
 })